+++ src/Actions/AlertStatusActions.php
@@ -1,0 +1,106 @@
+<?php
+namespace Chargebee\Actions;
+
+use Chargebee\Responses\AlertStatusResponse\AlertStatusesForAlertAlertStatusResponse;
+use Chargebee\Responses\AlertStatusResponse\AlertStatusesForSubscriptionAlertStatusResponse;
+use Chargebee\ValueObjects\Encoders\ListParamEncoder;
+use Chargebee\Actions\Contracts\AlertStatusActionsInterface;
+use Chargebee\ValueObjects\Encoders\URLFormEncoder;
+use Chargebee\ValueObjects\Transporters\ChargebeePayload;
+use Chargebee\ValueObjects\APIRequester;
+use Chargebee\HttpClient\HttpClientFactory;
+use Chargebee\Environment;
+use Exception;
+use Chargebee\Exceptions\PaymentException;
+use Chargebee\Exceptions\OperationFailedException;
+use Chargebee\Exceptions\APIError;
+use Chargebee\Exceptions\InvalidRequestException;
+
+final class AlertStatusActions implements AlertStatusActionsInterface
+{
+    private HttpClientFactory $httpClientFactory;
+    private Environment $env;
+    public function __construct(HttpClientFactory $httpClientFactory, Environment $env){
+       $this->httpClientFactory = $httpClientFactory;
+       $this->env = $env;
+    }
+
+    /**
+    *   @see https://apidocs.chargebee.com/docs/api/alert_statuses/list-alert-statuses-for-a-subscription?lang=php-v4
+    *   @param array{
+    *     limit?: int,
+    *     offset?: string,
+    *     alarm_status?: array{
+    *     is?: mixed,
+    *     },
+    * alert_id?: array{
+    *     in?: mixed,
+    *     },
+    * } $params Description of the parameters
+    *   @param string $id  
+    *   @param array<string, string> $headers
+    *   @return AlertStatusesForSubscriptionAlertStatusResponse
+    *   @throws PaymentException
+    *   @throws OperationFailedException
+    *   @throws APIError
+    *   @throws InvalidRequestException
+    *   @throws Exception
+    */
+    public function alertStatusesForSubscription(string $id, array $params = [], array $headers = []): AlertStatusesForSubscriptionAlertStatusResponse
+    {
+        $jsonKeys = [
+        ];
+        $payload = ChargebeePayload::builder()
+        ->withEnvironment($this->env)
+        ->withHttpMethod("get")
+        ->withUriPaths(["subscriptions",$id,"alert_statuses"])
+        ->withParamEncoder(new ListParamEncoder())
+        ->withSubDomain(null)
+        ->withJsonKeys($jsonKeys)
+        ->withHeaders($headers)
+        ->withParams($params)
+        ->build();
+        $apiRequester = new APIRequester($this->httpClientFactory, $this->env);
+        $respObject = $apiRequester->makeRequest($payload);
+        return AlertStatusesForSubscriptionAlertStatusResponse::from($respObject->data, $respObject->headers);
+    }
+
+    /**
+    *   @see https://apidocs.chargebee.com/docs/api/alert_statuses/list-alert-statuses-for-an-alert?lang=php-v4
+    *   @param array{
+    *     limit?: int,
+    *     offset?: string,
+    *     alarm_status?: array{
+    *     is?: mixed,
+    *     },
+    * } $params Description of the parameters
+    *   @param string $id  
+    *   @param array<string, string> $headers
+    *   @return AlertStatusesForAlertAlertStatusResponse
+    *   @throws PaymentException
+    *   @throws OperationFailedException
+    *   @throws APIError
+    *   @throws InvalidRequestException
+    *   @throws Exception
+    */
+    public function alertStatusesForAlert(string $id, array $params = [], array $headers = []): AlertStatusesForAlertAlertStatusResponse
+    {
+        $jsonKeys = [
+        ];
+        $payload = ChargebeePayload::builder()
+        ->withEnvironment($this->env)
+        ->withHttpMethod("get")
+        ->withUriPaths(["alerts",$id,"alert_statuses"])
+        ->withParamEncoder(new ListParamEncoder())
+        ->withSubDomain(null)
+        ->withJsonKeys($jsonKeys)
+        ->withHeaders($headers)
+        ->withParams($params)
+        ->build();
+        $apiRequester = new APIRequester($this->httpClientFactory, $this->env);
+        $respObject = $apiRequester->makeRequest($payload);
+        return AlertStatusesForAlertAlertStatusResponse::from($respObject->data, $respObject->headers);
+    }
+
+}
+?>