+++ src/Actions/Contracts/InvoiceActionsInterface.php
@@ -452,6 +452,9 @@ Interface InvoiceActionsInterface
     *     asc?: string,
     *     desc?: string,
     *     },
+    * exclude?: array{
+    *     in?: mixed,
+    *     },
     * } $params Description of the parameters
     *   
     *   @param array<string, string> $headers
@@ -1238,13 +1241,13 @@ Interface InvoiceActionsInterface
     *     offset?: string,
     *     payment_reference_number?: array{
     *     number?: array{
-    *         is?: string,
-    *             in?: string,
+    *         in?: string,
+    *             is?: string,
     *             },
     *     },
     * id?: array{
-    *     is?: mixed,
     *     in?: mixed,
+    *     is?: mixed,
     *     },
     * } $params Description of the parameters
     *