+++ src/Actions/Contracts/ItemActionsInterface.php
@@ -96,8 +96,8 @@ Interface ItemActionsInterface
     *     not_in?: mixed,
     *     },
     * business_entity_id?: array{
-    *     is?: mixed,
     *     is_present?: mixed,
+    *     is?: mixed,
     *     },
     * include_site_level_resources?: array{
     *     is?: mixed,