+++ src/Actions/Contracts/VirtualBankAccountActionsInterface.php
@@ -86,6 +86,7 @@ Interface VirtualBankAccountActionsInterface
     *   @param array{
     *     customer_id?: string,
     *     email?: string,
+    *     gateway_account_id?: string,
     *     scheme?: string,
     *     } $params Description of the parameters
     *   
@@ -132,6 +133,7 @@ Interface VirtualBankAccountActionsInterface
     *   @param array{
     *     customer_id?: string,
     *     reference_id?: string,
+    *     gateway_account_id?: string,
     *     scheme?: string,
     *     } $params Description of the parameters
     *