+++ src/Actions/ExportActions.php
@@ -479,8 +479,8 @@ final class ExportActions implements ExportActionsInterface
     *             },
     *     },
     * business_entity_id?: array{
-    *     is?: mixed,
     *     is_present?: mixed,
+    *     is?: mixed,
     *     },
     * include_site_level_resources?: array{
     *     is?: mixed,
@@ -738,8 +738,8 @@ final class ExportActions implements ExportActionsInterface
     *             },
     *     },
     * business_entity_id?: array{
-    *     is?: mixed,
     *     is_present?: mixed,
+    *     is?: mixed,
     *     },
     * include_site_level_resources?: array{
     *     is?: mixed,
@@ -847,8 +847,8 @@ final class ExportActions implements ExportActionsInterface
     *             },
     *     },
     * business_entity_id?: array{
-    *     is?: mixed,
     *     is_present?: mixed,
+    *     is?: mixed,
     *     },
     * include_site_level_resources?: array{
     *     is?: mixed,
@@ -1828,8 +1828,8 @@ final class ExportActions implements ExportActionsInterface
     *     not_in?: mixed,
     *     },
     * applicable_item_price_ids?: array{
-    *     is?: mixed,
     *     in?: mixed,
+    *     is?: mixed,
     *     },
     * } $params Description of the parameters
     *   
@@ -2124,8 +2124,8 @@ final class ExportActions implements ExportActionsInterface
     *     not_in?: mixed,
     *     },
     * business_entity_id?: array{
-    *     is?: mixed,
     *     is_present?: mixed,
+    *     is?: mixed,
     *     },
     * include_site_level_resources?: array{
     *     is?: mixed,