+++ src/Actions/InvoiceActions.php
@@ -696,6 +696,9 @@ final class InvoiceActions implements InvoiceActionsInterface
     *     asc?: string,
     *     desc?: string,
     *     },
+    * exclude?: array{
+    *     in?: mixed,
+    *     },
     * } $params Description of the parameters
     *   
     *   @param array<string, string> $headers
@@ -1770,13 +1773,13 @@ final class InvoiceActions implements InvoiceActionsInterface
     *     offset?: string,
     *     payment_reference_number?: array{
     *     number?: array{
-    *         is?: string,
-    *             in?: string,
+    *         in?: string,
+    *             is?: string,
     *             },
     *     },
     * id?: array{
-    *     is?: mixed,
     *     in?: mixed,
+    *     is?: mixed,
     *     },
     * } $params Description of the parameters
     *