+++ src/Actions/ItemFamilyActions.php
@@ -83,8 +83,8 @@ final class ItemFamilyActions implements ItemFamilyActionsInterface
     *     between?: mixed,
     *     },
     * business_entity_id?: array{
-    *     is?: mixed,
     *     is_present?: mixed,
+    *     is?: mixed,
     *     },
     * include_site_level_resources?: array{
     *     is?: mixed,