+++ src/Resources/Card/Enums/CardType.php
@@ -23,6 +23,7 @@ enum CardType : string {
     case MAESTRO = "maestro";
     case DANKORT = "dankort";
     case CARTES_BANCAIRES = "cartes_bancaires";
+    case MADA = "mada";
     case OTHER = "other";
     case NOT_APPLICABLE = "not_applicable";
     case UNKNOWN = "unknown";