+++ src/Responses/AlertStatusResponse/AlertStatusesForSubscriptionAlertStatusResponseListObject.php
@@ -1,0 +1,14 @@
+<?php
+namespace Chargebee\Responses\AlertStatusResponse;
+
+use Chargebee\Resources\AlertStatus\AlertStatus;
+
+class AlertStatusesForSubscriptionAlertStatusResponseListObject
+{ 
+    public AlertStatus $alert_status;
+    public function __construct(
+        AlertStatus $alert_status,
+    ) { 
+        $this->alert_status = $alert_status;
+    }
+}